+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/ColumnPsiReference.kt
@@ -1,6 +1,6 @@
 package dev.ekvedaras.laravelquery.reference
 
-import com.intellij.database.psi.DbPsiFacade
+import com.intellij.database.util.DbUtil
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReferenceBase
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
@@ -12,8 +12,13 @@ class ColumnPsiReference(element: PsiElement) : PsiReferenceBase<PsiElement>(ele
 
         rangeInElement = target.ranges.last()
 
-        return DbPsiFacade.getInstance(element.project).findElement(
-            target.column.find { tables.contains(it.tableName) }
-        )
+        DbUtil.getDataSources(element.project).forEach { dataSource ->
+            val dbColumn = dataSource.findElement(target.column.find { tables.contains(it.tableName) })
+            if (dbColumn != null) {
+                return dbColumn
+            }
+        }
+
+        return null
     }
 }