+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/ForeignKeyPsiReference.kt
@@ -1,6 +1,6 @@
 package dev.ekvedaras.laravelquery.reference
 
-import com.intellij.database.psi.DbPsiFacade
+import com.intellij.database.util.DbUtil
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReferenceBase
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
@@ -12,8 +12,13 @@ class ForeignKeyPsiReference(element: PsiElement) : PsiReferenceBase<PsiElement>
 
         rangeInElement = target.ranges.last()
 
-        return DbPsiFacade.getInstance(element.project).findElement(
-            target.foreignKey.find { tables.contains(it.table?.name) }
-        )
+        DbUtil.getDataSources(element.project).forEach { dataSource ->
+            val dbForeignKey = dataSource.findElement(target.foreignKey.find { tables.contains(it.table?.name) })
+            if (dbForeignKey != null) {
+                return dbForeignKey
+            }
+        }
+
+        return null
     }
 }