+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/TableOrViewPsiReference.kt
@@ -1,6 +1,6 @@
 package dev.ekvedaras.laravelquery.reference
 
-import com.intellij.database.psi.DbPsiFacade
+import com.intellij.database.util.DbUtil
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReferenceBase
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
@@ -16,6 +16,13 @@ class TableOrViewPsiReference(element: PsiElement, private val type: DbReference
             target.ranges.first()
         }
 
-        return DbPsiFacade.getInstance(element.project).findElement(target.table.firstOrNull())
+        DbUtil.getDataSources(element.project).forEach { dataSource ->
+            val dbTable = dataSource.findElement(target.table.firstOrNull())
+            if (dbTable != null) {
+                return dbTable
+            }
+        }
+
+        return null
     }
 }